Rename url package references to uri
