+++ library/src/main/java/dev/gerlot/securewebview/uri/BeginningMatcher.java
@@ -1,20 +1,20 @@
-package dev.gerlot.securewebview.url;
+package dev.gerlot.securewebview.uri;
 
 import android.net.Uri;
 
 import androidx.annotation.NonNull;
 
-public class ExactUriMatcher implements UriMatcher {
+public class BeginningMatcher implements UriMatcher {
 
     private final Uri mUri;
 
-    public ExactUriMatcher(Uri uri) {
+    public BeginningMatcher(Uri uri) {
         mUri = uri;
     }
 
     @Override
     public boolean matches(@NonNull Uri uri) {
-        return mUri.equals(uri);
+        return uri.toString().startsWith(mUri.toString());
     }
 
 }
