+++ library/src/main/java/dev/gerlot/securewebview/uri/UriMatcher.java
@@ -1,4 +1,4 @@
-package dev.gerlot.securewebview.url;
+package dev.gerlot.securewebview.uri;
 
 import android.net.Uri;
