+++ library/src/test/java/dev/gerlot/securewebview/uri/AuthorityWithWwwMatcherTest.java
@@ -1,4 +1,4 @@
-package dev.gerlot.securewebview.url;
+package dev.gerlot.securewebview.uri;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -10,10 +10,10 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 @RunWith(RobolectricTestRunner.class)
-public class AuthorityContainmentMatcherTest {
+public class AuthorityWithWwwMatcherTest {
 
     public static final Uri GOOGLE = new Uri.Builder().authority("google.com").appendPath("search").build();
-    private static final UriMatcher MATCHER = new AuthorityContainmentMatcher(GOOGLE);
+    private static final UriMatcher MATCHER = new AuthorityWithWwwMatcher(GOOGLE);
 
     @Test
     public void testMatches_same() {
@@ -33,6 +33,11 @@ public class AuthorityContainmentMatcherTest {
     @Test
     public void testMatches_differentPath() {
         assertTrue(MATCHER.matches(new Uri.Builder().authority("google.com").appendPath("post").build()));
+    }
+
+    @Test
+    public void testMatches_wwwPrefix() {
+        assertFalse(MATCHER.matches(new Uri.Builder().authority("www.google.com").build()));
     }
 
 }
