+++ library/src/test/java/dev/gerlot/securewebview/uri/AuthorityWithoutWwwMatcherTest.java
@@ -1,7 +1,6 @@
-package dev.gerlot.securewebview.url;
+package dev.gerlot.securewebview.uri;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import android.net.Uri;
 
@@ -10,10 +9,10 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 @RunWith(RobolectricTestRunner.class)
-public class ExactUriMatcherTest {
+public class AuthorityWithoutWwwMatcherTest {
 
     public static final Uri GOOGLE = new Uri.Builder().authority("google.com").appendPath("search").build();
-    private static final UriMatcher MATCHER = new ExactUriMatcher(GOOGLE);
+    private static final UriMatcher MATCHER = new AuthorityWithoutWwwMatcher(GOOGLE);
 
     @Test
     public void testMatches_same() {
@@ -32,7 +31,12 @@ public class ExactUriMatcherTest {
 
     @Test
     public void testMatches_differentPath() {
-        assertFalse(MATCHER.matches(new Uri.Builder().authority("google.com").appendPath("post").build()));
+        assertTrue(MATCHER.matches(new Uri.Builder().authority("google.com").appendPath("post").build()));
+    }
+
+    @Test
+    public void testMatches_wwwPrefix() {
+        assertTrue(MATCHER.matches(new Uri.Builder().authority("www.google.com").build()));
     }
 
 }
