+++ library/src/test/java/dev/gerlot/securewebview/uri/DeniedUriListTest.java
@@ -1,4 +1,4 @@
-package dev.gerlot.securewebview.url;
+package dev.gerlot.securewebview.uri;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
